requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
openai
psycopg2-binary
celery
redis
python-crontab
schedule
